Move all components inside features folder
