--- src/components/chart/index.js
@@ -1,1 +1,0 @@
-export { Chart } from "./Chart";
