+++ src/features/chart/components/index.js
@@ -1,0 +1,1 @@
+export { Chart } from "./Chart";
