+++ src/features/headline/components/adaptive-logo.png
