+++ src/features/key-stats/components/KeyStats.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { useSelector } from "react-redux";
-import { Loading } from "../loading";
+import { Loading } from "../../loading/components/Loading";
 import "./KeyStats.css";
 
 const NUMBER_FORMATTER = new Intl.NumberFormat();
