+++ src/features/stock-ticker/components/StockTicker.css
