+++ src/features/stock-ticker/components/StockTicker.js
