+++ src/features/stock-ticker/components/index.js
